requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask-bcrypt
flask-jwt-extended
flask-login
pillow
opencv-python
pytz